clean app.jsx add refresh user moved App.jsx to components folder changed paths to components
