+++ src/components/App/App.jsx
@@ -1,0 +1,73 @@
+import { Suspense, lazy, useEffect } from "react";
+import { useDispatch } from "react-redux";
+import { Route, Routes } from "react-router-dom";
+
+import { refreshUser } from "../../redux/auth/operations.js";
+
+import SharedLayout from "../../components/SharedLayout/SharedLayout";
+import RestrictedRoute from "../../components/Route/RegisteredRoute/RegisteredRoute.jsx";
+import PrivateRoute from "../../components/Route/PrivateRout/PrivateRout.jsx";
+import Loader from "../../components/Loader/Loader.jsx";
+import { useAuth } from "../../hooks/userAuth.js";
+
+const WelcomePage = lazy(() =>
+  import("../../pages/WelcomePage/WelcomePage.jsx")
+);
+const SignupPage = lazy(() => import("../../pages/SignupPage/SignupPage.jsx"));
+const SigninPage = lazy(() => import("../../pages/SigninPage/SigninPage.jsx"));
+const HomePage = lazy(() => import("../../pages/HomePage/HomePage.jsx"));
+const NotFoundPage = lazy(() =>
+  import("../../pages/NotFoundPage/NotFoundPage.jsx")
+);
+
+function App() {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    if (useAuth.isLoggedIn) {
+      dispatch(refreshUser());
+    }
+    return;
+  }, [dispatch]);
+
+  return useAuth.isRefreshing ? (
+    <b>Refreshing user...</b>
+  ) : (
+    <>
+      <Suspense fallback={<Loader />}>
+        <Routes>
+          <Route path="/" element={<SharedLayout />}>
+            <Route path="/welcome" element={<WelcomePage />} />
+            <Route
+              path="/signup"
+              element={
+                <RestrictedRoute
+                  redirectTo="/home"
+                  component={<SignupPage />}
+                />
+              }
+            />
+            <Route
+              path="/signin"
+              element={
+                <RestrictedRoute
+                  redirectTo="/home"
+                  component={<SigninPage />}
+                />
+              }
+            />
+            <Route
+              path="/home"
+              element={
+                <PrivateRoute redirectTo="/signin" component={<HomePage />} />
+              }
+            />
+            <Route path="*" element={<NotFoundPage />} />
+          </Route>
+        </Routes>
+      </Suspense>
+    </>
+  );
+}
+
+export default App;
