+++ src/main.jsx
@@ -1,6 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import App from "./App.jsx";
+import App from "./components/App/App.jsx";
 import "./index.css";
 import "modern-normalize";
 // import { PersistGate } from "redux-persist/integration/react";
